frontend added and bugs fixed

+++ catalog/views.py
@@ -1,5 +1,8 @@
 from django.shortcuts import render, redirect
 from . import models
+import telebot
+
+bot = telebot.TeleBot("6269391112:AAFlxxKYVCaFS6l2BS26nmMATi3bSCoQ1hg")
 from django.http import HttpResponse
 # Create your views here.
 
@@ -40,10 +43,20 @@ def get_cart(request):
 def complete_order(request):
     user_cart=models.UserCart.objects.filter(user_id=request.user.id)
     # Сообщение в тг для админа
-    result_message = "Новый заказ(из Сайта)\n\n"
-    for cart in user_cart:
-        result_message += f"{cart.user_product}\n" \
-                          f"Количество: {cart.user_product_quantity}"
-    bot.send_message(305896408, result_message)
-    user_cart.delete()
-    return redirect("/")
+
+    if request.method == "POST":
+        result_message = "Новый заказ(из Сайта)\n\n"
+        total = 0
+        for cart in user_cart:
+            result_message += f"Название продукта: {cart.user_product}\n"\
+                          f"Количество: {cart.user_product_quantity}\n\n"
+            total += cart.user_product.price * cart.user_product_quantity
+        result_message += f"\n\nИтог: {total}"
+        bot.send_message(305896408, result_message)
+        user_cart.delete()
+        return redirect("/")
+    return render(request, "user_cart.html", {"user_cart": user_cart})
+def delete_from_user_cart(request, pk):
+    product_to_delete = models.Product.objects.get(id=pk)
+    models.UserCart.objects.filter(user_id=request.user.id, user_product=product_to_delete).delete()
+    return redirect('/cart')
